azure-function/requirements.txt
azure-functions==1.18.0
azure-storage-queue==12.8.0
requests==2.31.0
pydantic==2.5.0